+++ packages/odyssey-react-mui/src/Field.tsx
@@ -30,6 +30,14 @@ export const fieldTypeValues = ["single", "group"] as const;
 
 export type FieldProps = {
   /**
+   * If `error` is not undefined, the `input` will indicate an error.
+   */
+  errorMessage?: string;
+  /**
+   * If `error` is not undefined, the `input` will indicate an error.
+   */
+  errorMessageList?: string[];
+  /**
    * The field type determines how ARIA components are setup. It's important to use this to denote if you expect only one component (like a text field) or multiple (like a radio group).
    */
   fieldType: (typeof fieldTypeValues)[number];
@@ -72,7 +80,9 @@ export type FieldProps = {
 };
 
 const Field = ({
+  ariaDescribedBy,
   errorMessage,
+  errorMessageList,
   fieldType,
   hasVisibleLabel,
   hint,
@@ -87,7 +97,9 @@ const Field = ({
 }: FieldProps &
   Pick<
     FieldComponentProps,
+    | "ariaDescribedBy"
     | "errorMessage"
+    | "errorMessageList"
     | "hint"
     | "HintLinkComponent"
     | "id"
@@ -99,12 +111,15 @@ const Field = ({
 
   const id = useUniqueId(idOverride);
   const hintId = hint ? `${id}-hint` : undefined;
-  const errorMessageElementId = errorMessage ? `${id}-error` : undefined;
+  const errorMessageElementId =
+    errorMessage || errorMessageList ? `${id}-error` : undefined;
   const labelElementId = `${id}-label`;
 
-  const ariaDescribedBy = useMemo(
-    () => [hintId, errorMessageElementId].join(" ").trim() || undefined,
-    [errorMessageElementId, hintId]
+  const localAriaDescribedBy = useMemo(
+    () =>
+      [hintId, errorMessageElementId, ariaDescribedBy].join(" ").trim() ||
+      undefined,
+    [ariaDescribedBy, errorMessageElementId, hintId]
   );
 
   const { isDisabled: isFieldsetDisabled } = useFieldset();
@@ -118,7 +133,10 @@ const Field = ({
     <MuiFormControl
       component={fieldType === "group" ? "fieldset" : "div"}
       disabled={isDisabled}
-      error={Boolean(errorMessage)}
+      error={
+        Boolean(errorMessage) ||
+        (Array.isArray(errorMessageList) && errorMessageList.length > 0)
+      }
       role={isRadioGroup ? "radiogroup" : undefined}
       fullWidth={isFullWidth}
     >
@@ -146,14 +164,18 @@ const Field = ({
       )}
 
       {renderFieldComponent({
-        ariaDescribedBy,
+        ariaDescribedBy: localAriaDescribedBy,
         errorMessageElementId,
         id,
         labelElementId,
       })}
 
-      {errorMessage && (
-        <FieldError id={errorMessageElementId} text={errorMessage} />
+      {(errorMessage || errorMessageList) && (
+        <FieldError
+          id={errorMessageElementId}
+          message={errorMessage}
+          messageList={errorMessageList}
+        />
       )}
     </MuiFormControl>
   );